Async into. Replaces async toArray
