+++ async/into.js
@@ -1,0 +1,2 @@
+'use strict'
+module.exports = require('../core/_coreInto')(require('./_core'))
