+++ core/into.js
@@ -1,43 +1,2 @@
 'use strict'
-var core = require('./_core'),
-    transduce = core._transduce,
-    reduce = core._reduce,
-    transformer = require('./transformer'),
-    isFunction = require('./util').isFunction
-
-module.exports =
-function into(init, t, coll){
-  var xf = transformer(init),
-      len = arguments.length
-
-  if(len === 1){
-    return intoCurryXf(xf)
-  }
-
-  if(len === 2){
-    if(isFunction(t)){
-      return intoCurryXfT(xf, t)
-    }
-    coll = t
-    return reduce(xf, init, coll)
-  }
-  return transduce(t, xf, init, coll)
-}
-
-function intoCurryXf(xf){
-  return function intoXf(t, coll){
-    if(arguments.length === 1){
-      if(isFunction(t)){
-        return intoCurryXfT(xf, t)
-      }
-      coll = t
-      return reduce(xf, xf.init(), coll)
-    }
-    return transduce(t, xf, xf.init(), coll)
-  }
-}
-function intoCurryXfT(xf, t){
-  return function intoXfT(coll){
-    return transduce(t, xf, xf.init(), coll)
-  }
-}
+module.exports = require('./_coreInto')(require('./_core'))
